+++ src/core/endorsement-chain/index.ts
@@ -1,6 +1,7 @@
 export * from './fetchEscrowTransfer';
 export * from './fetchTokenTransfer';
 export * from './helpers';
+export * from './obligation';
 export * from './retrieveEndorsementChain';
 export * from './types';
 export * from './useEndorsementChain';